Enabling Xcode 6.1 Compiler Options and Removing warning
New compiler options for xCode 6.1 were added
A compiler warning was removed in TSActionSheet.m

+++ TSPopover/TSActionSheet.m
@@ -199,7 +199,7 @@
 - (void)buttonClicked:(id)sender 
 {
     /* Run the button's block */
-    int buttonIndex = [sender tag] - 1;
+    NSInteger buttonIndex = [sender tag] - 1;
     [self dismissWithClickedButtonIndex:buttonIndex animated:YES];
 }
